intelligent-data-room-be/requirements.txt
fastapi[all]
python-dotenv
google-genai